backend/requirements.txt
fastapi
uvicorn
motor
python-dotenv
pydantic[email]
PyJWT
